refactor: Optimised codes.
Move all private or protected to the last of files.
